fix timeseries chart
