+++ src/dpr/components/_dashboards/dashboard-filters/style.scss
@@ -6,6 +6,7 @@
     position: sticky;
     top: 0;
     z-index: 1000;
+    border-bottom: solid white govuk-spacing(2);
     background-color: govuk-colour("light-grey");
 
       .dpr-interactive-filters {
