+++ src/dpr/components/_dashboards/dashboard-section/utils.ts
@@ -5,7 +5,7 @@ const getDatasetRows = (listDefinition: DashboardVisualisation, dashboardData: M
   const { keys, measures, ignore, values } = listDefinition.columns
 
   const displayColumnsIds = measures.map((col) => col.id)
-  const keyColumnsIds = keys.map((col) => col.id)
+  const keyColumnsIds = keys?.map((col) => col.id) || []
   const ignoreIds = ignore?.map((col) => col.id) || []
   const valueIds = values?.map((col) => col.id) || []
   const expectAllNoneDefinedColsToBeNull = ignore && ignore.length === 0
@@ -15,19 +15,16 @@ const getDatasetRows = (listDefinition: DashboardVisualisation, dashboardData: M
   return dashboardData.filter((datasetRow: MetricsDataResponse) => {
     const validRow: boolean[] = []
     Object.keys(datasetRow).forEach((datasetField) => {
-      // console.log({ datasetField })
-
       const value = datasetRow[datasetField].raw
-
       // All rows are valid until proven otherwise
       let valid = true
 
-      // 2. check if the column value is equal to a defined column value
+      // 1. check if the column value is equal to a defined column value
       if (valueIds.includes(datasetField)) {
         const valueColumn = values.find((col) => col.id === datasetField)
         valid = value === valueColumn.equals
 
-        // 1. check values exist in the defined columns
+        // 2. check values exist in the defined columns
       } else if (displayColumnsIds.includes(datasetField) || keyColumnsIds.includes(datasetField)) {
         valid = value !== '' && value !== undefined && value !== null
